moved DR reduction to preprocessing, added more functionality for bitvec expressions
